chore: renaming file to api.ts because route.ts would need a GET() handler
